# Performance improvements in GetRateLimits()