## Purpose
To incorporate the performance improvements recommended in #74 Comments by @valer-cara 

## Implementation
* GetRateLimits() now processes requests synchronously until the key is found to belong to a remote instance. A go-routine is then spawned to handle the remote call. 

## Results
While the change did show an improvement, I didn't see the huge improvement I was expecting in `GetRateLimit` tests. The biggest improvement is in the `ThunderingHerd` tests where there is almost a 1x improvement. Users should notice the greatest improvement for rate limits when using `Behavior = GLOBAL` and making many concurrent requests.

```
== Master ==
BenchmarkServer_GetPeerRateLimitNoBatching
BenchmarkServer_GetPeerRateLimitNoBatching/GetPeerRateLimitNoBatching
BenchmarkServer_GetPeerRateLimitNoBatching/GetPeerRateLimitNoBatching-12            10000    108844 ns/op
BenchmarkServer_GetRateLimit
BenchmarkServer_GetRateLimit/GetRateLimit
BenchmarkServer_GetRateLimit/GetRateLimit-12                                         1581    778386 ns/op
BenchmarkServer_GetRateLimitGlobal
BenchmarkServer_GetRateLimitGlobal/GetRateLimitGlobal
BenchmarkServer_GetRateLimitGlobal/GetRateLimitGlobal-12                             4720    248670 ns/op
BenchmarkServer_Ping
BenchmarkServer_Ping/HealthCheck
BenchmarkServer_Ping/HealthCheck-12                                                 13084     90224 ns/op
BenchmarkServer_ThunderingHeard
BenchmarkServer_ThunderingHeard/ThunderingHeard
BenchmarkServer_ThunderingHeard/ThunderingHeard-12                                  45579     26908 ns/op

== Performance Improvements ==
BenchmarkServer_GetPeerRateLimitNoBatching
BenchmarkServer_GetPeerRateLimitNoBatching/GetPeerRateLimitNoBatching
BenchmarkServer_GetPeerRateLimitNoBatching/GetPeerRateLimitNoBatching-12            11348    102626 ns/op
BenchmarkServer_GetRateLimit
BenchmarkServer_GetRateLimit/GetRateLimit
BenchmarkServer_GetRateLimit/GetRateLimit-12                                         1612    739589 ns/op
BenchmarkServer_GetRateLimitGlobal
BenchmarkServer_GetRateLimitGlobal/GetRateLimitGlobal
BenchmarkServer_GetRateLimitGlobal/GetRateLimitGlobal-12                             5185    199153 ns/op
BenchmarkServer_Ping
BenchmarkServer_Ping/HealthCheck
BenchmarkServer_Ping/HealthCheck-12                                                 13460     89366 ns/op
BenchmarkServer_ThunderingHerd
BenchmarkServer_ThunderingHerd/ThunderingHerd
BenchmarkServer_ThunderingHerd/ThunderingHerd-12                                    76543     16422 ns/op
```